tiny update, we can use lab module type enum, which was stored as int before, now as enum, stepping harder a little while maintining compatability between tcc and our language

+++ compiler/lab/LabBuildContext.cpp
@@ -74,7 +74,7 @@ LabModule *LabBuildContext::add_with_type(
     LabModule **dependencies,
     unsigned int dep_len
 ) {
-    modules.emplace_back((int) type, std::move(name));
+    modules.emplace_back(type, std::move(name));
     auto mod = &modules.back();
     LabBuildContext::add_paths(mod->paths, paths, path_len);
     LabBuildContext::add_dependencies(mod->dependencies, dependencies, dep_len);
